version 2.3.0
 - remove TIMER flag and enable the timer in all builds.
 - quadruple the number of trians in nrun profiling.
 - change bwrn profiling trial count from 24 to 64.
 - make `keypressed` macro instead of using GetAsyncKeyState
 - add `-R` which sets as REALTIME priority and sets the CPU affinity mask.
 - fix `-v` for HELP=false. previously it stored "life vx.y.z" and "x.y.z" as separate strings.
 - change printf formatting: 0x%...x => %#...x, %llu => %zu, %llx => %zx, etc.
 - update data.json
 - allow stop keypresses in the middle of predecessor searches.
 - disallow the update key in bwrn. instead, print the state and state number every time.
 - implement the timer for bwrn.
 - use 32-bit OOM codes instead of 64-bit codes.
 - change summary.h to be something like stb where the prototype and implementation are separate.
 - change the file output code.
    - change from 5 tries per second to 3.
    - lock at byte 0 instead of at the end. only lseek to the end after the lock is acquired.

+++ include/errlog.h
@@ -80,12 +80,12 @@
 #define enputs(str) enfputs(stderr, str)
 #define edputs(str) edfputs(stderr, str)
 
-#define OOM(mem, code) ({                    \
-	if ((mem) == NULL) {                      \
-		eprintf("Out of Memory. code: %llu\n", \
-			(unsigned long long int) (code));   \
-		exit(ERRLOG_OOM_EC);                     \
-	}                                             \
+#define OOM(mem, code) ({                  \
+	if ((mem) == NULL) {                    \
+		eprintf("Out of Memory. code: %u\n", \
+			(unsigned int) (code));           \
+		exit(ERRLOG_OOM_EC);                   \
+	}                                           \
 })
 
 #define _VA_ID_IGNORED(...)
